cleaned some stuff up probably moved some of your stuff around sorry
Even though we are hacking we do need to just clean up the  cruft that builds up in these files over time  so many of these files had never really been touched in months and were just sitting in the top directory causing confusion. If you need to go back in reference on file you had git the commit hash and look it up on github.